threeStyleQuizLogに関するAPIを別ファイルに切り出した
